remove tests/sigtran: it's not a test case
in tests/* we have unit tests that are run as part of the autotest suite
during 'make check'.  The code in tests/sigtran is an example, but not
a test.  As the API is changing anyway, let's remove it for now and
re-introduce actual tests and examples after the changes in API required
by the upcoming new SCCP core.

Change-Id: Ie471a197856c875eb4987bf9858d757312de24fb