sua2sccp: Only encode SCCP options permitted for given msg type
The SCCP spec clearly defines which optional information element each of
the messages supports.  We must make sure to not include any other
options when converting from SUA.  SUA has more relaxed rules about
this, and e.g. supports SRC and DEST ADDR in the COAK, while the SCCP CC
only permits a CalledParty, but not CallingParty.

This was found when interoperating against the Ericsson TTCN-3 SCCP
implementation for Eclipse Titan.

Change-Id: I7d9e23dec1d3e786d291abd270a261704593befc