sccp: check for valid point code in osmo_sccp_simple_client()
The point-code is not checked in the beginning, only the ASP
checks the point code at a very late stage and fails in case
an invalid point-code is detected.

Remove the check in the ASP creation.

Add a new check after the SS7 initalization that checks the
point-code. If none is set, the default point-code will be
used.

Change-Id: I334d90e769bd9952f67c51b12a945f22bd268fa8