ipa: Patch DPC/OPC information into SCCP Calling/Called Party Address
When we receive a SCCPlite message from an IPA peer, it may simply
contain a SSN number but no point codes.  Similarly to creating a fake
MTP routing label from override DPC and routing key OPC, we can also
add that point code information to the SCCP header.  This way the rest
of the SS7 network can handle the message and route it accordingly.

Change-Id: I4a2ff063e3c060641b3fd181a1cd600da3ec568b