contrib/test: Fix config of local/remote ports
The first port is the remote one, and the second port is the local one,
according to cs7_asp_cmd doc and code. In the same config, the ports for
the servers are used and for the local port in the client we don't care,
that's why we use 0 there.

Change-Id: I0fafd07614068a27c19bc2dfa6491b4b0c6737fb