sccp_helpers: don't return msgb with l2h set
It's a bad idea to use sccp_msgb_alloc() for SCCP User Primitive msgbs.

The rationale is quite simple: The SCU msgb's are used for wrapping
osmo_prim.  The user payload data (e.g. BSSAP) in such primitives is
found at msgb->l2h.  However, user payload data is optional.  So in a
SCU primitive without user data, we must have msgb->l2h == NULL.

The old behavior resulted in bogus data (actually the sccp_user_prim)
to be contained in the DATA section of SCCP messages such as RLSD/RLC.

Also, the old implementation of scu_msgb_alloc() discarded the 'name'
argument and replaced it with a static "SCU" which was of course another
bug.

Change-Id: I19cb83302aaa404ab1a2d92e6f2aec43d0380426
Related: OS#2732