sccp_types.h: Fix value for SCCP_REFUSAL_UNEQUIPPED_USER
It seems we have been sending the wrong numeric value in SCCP connection
refusal due to an unqeuipped user.  It turns out our list of refusal
causes was missing one entry, causing an off-by-one for this refusal
cause.  While at it, add a comment which section of which spec is
relevant for this enum.

Change-Id: I113645bd6df1ec9ae5137977028df38560fc4789