xua: ipa_asp_fsm: Allow moving ASP to inactive state
If AS is configured with Traffic Mode Override, then if a new ASP
becomes active, all previous ASPs are considered to be inactive and new
data is sent to the newly activated ASP.

Remark: It's still unclear which methodology/implementation will follow
when the last activated ASP becomes inactive/shutdown. Then probably
another one should be activated at that time, but that logic is not
there implemented as far as I know.

Change-Id: I4ff246b2f899aaa3cf63bbdb3f3d317dc89b3d15