Fix (2) change in ss7 server default listen addr, keeps backward-compatibility behavior
This patch is a fixup or extension of commit
96d348efbf29888f5ef923251c7d51d45a897c0b, which only fixed the default
values for "asp" VTY node, but not for the xua server's "listen" node.
As a result, without this patch, by default the SCTP server socket will
only listen on IPv4's "0.0.0.0", since NULL is resolved by getaddrinfo
to it instead of "::" as first candidate.

Fixes: 96d348efbf29888f5ef923251c7d51d45a897c0b
Change-Id: Ifbc8df854d1f1e9b07b11911ad5da8cdf9f2080a