HACK: Add hard-coded DPC to SSA in response to SST
At least some version of an Ericsson MSC is sending us a SST with
point codes only in the M3UA header, but not in the SCCP Called/Calling
address (which only contain SSN).  So if we create a SSA in respone, it
will only have SSN in it, and hence it cannot be routed.

We hard-code the destination point code (7000) here just for testing.

A proper solution would have to either
a) set up some SCCP routing that routes the SSN to a given point code
b) add the M3UA point codes from the MTP/M3UA layer to the SCCP
   Calling/CalledParty when receiving the message at the M3UA->SCCP boundary