ASPAC/ASPIA: Don't add routing context IE in ASP-role if routing context 0
If there is only one routing key (and hence routing context, and hence
AS) within one ASP, *and* the routing context is configured as 0 (zero),
we should not include the routing context IE.  This is the way how
libosmo-sigtran has always been special-casing routing context 0 meaning
"routing context is not used".  However, that was only working in SG
role (typical STP use case).  When operating in ASP role, the
special-case handling was missing, causing a routing context IE
containing zero to be included in the related transmitted M3UA ASPAC
and ASPIA messages.

Change-Id: I5ce89b393a3b950ab7fd1eace9038718c9efcc51
Closes: OS#6003